+++ tests/testthat/test_format_dif.R
@@ -1,4 +1,6 @@
 context("DIF imports/exports")
-require("datasets")
 
-#test_that("Import from DIF", {})
+test_that("Import from DIF", {
+    dd <- import(file.path(system.file("misc", package = "utils"), "exDIF.dif"), transpose = TRUE)
+    expect_true(inherits(dd, "data.frame"))
+})