+++ tests/testthat/test_guess.R
@@ -1,4 +1,5 @@
 context("Get File Extension")
+library("datasets")
 
 test_that("File extension converted correctly", {
     expect_that(get_ext("hello.csv"), equals("csv"))
@@ -16,4 +17,9 @@ test_that("Format converted correctly", {
     expect_that(get_type("stata"), equals("dta"))
     expect_that(get_type("spss"), equals("sav"))
     expect_that(get_type("sas"), equals("sas7bdat"))
+})
+
+test_that("Export without file specified", {
+    expect_true(export(iris, format = "csv") %in% dir())
+    unlink("iris.csv")
 })